requirements.txt
PyGithub
graphviz